IOT Final Project-20260422T021050Z-3-001/IOT Final Project/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.0
python-dotenv==1.0.0
PyJWT==2.12.1
cryptography==41.0.3
requests==2.31.0
Werkzeug==2.3.7
PyMySQL==1.1.0
